backend_tienda/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
bcrypt==4.1.2
python-dotenv
pytest
pytest-asyncio
httpx
pytest-cov